.submodules/nostos/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nostos"
version = "0.1.0"
description = "Bring your bare metal home to the Talos cluster"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Yuri Lima" }]
keywords = ["talos", "kubernetes", "pxe", "bare-metal", "provisioning"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: System :: Installation/Setup",
]
dependencies = [
  "click>=8.1.0",
  "rich>=13.7.0",
  "questionary>=2.0.0",
  "httpx>=0.27.0",
  "pyyaml>=6.0",
  "pydantic>=2.6.0",
]

[project.optional-dependencies]
web = [
  "fastapi>=0.110.0",
  "uvicorn>=0.29.0",
  "jinja2>=3.1.0",
]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.4.0",
  "mypy>=1.9.0",
  "types-PyYAML>=6.0",
]

[project.scripts]
nostos = "nostos.cli:main"

[project.urls]
Repository = "https://github.com/yurifrl/nostos"
Issues = "https://github.com/yurifrl/nostos/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/nostos"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "C4"]
ignore = ["E501"] # line length handled by formatter

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
